+++ .gitignore
@@ -32,5 +32,6 @@
 *.app
 
 # standard build directories
-build
-_build
+/build*/
+/install*/
+/_*/
