+++ src/damping/dftd_rational.cpp
@@ -25,8 +25,8 @@
 #include "dftd_matrix.h"
 #include "dftd_ncoord.h"
 #include "dftd_parameters.h"
-#include "damping/atm.h"
-#include "damping/rational.h"
+#include "damping/dftd_atm.h"
+#include "damping/dftd_rational.h"
 
 
 namespace dftd4 {
